requirements.txt
# Using compatible versions to avoid dependency conflicts
solana-sdk==0.25.6
web3==6.11.1
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
aiohttp==3.9.1
base58==2.1.1
loguru==0.7.0
cryptography>=3.4.0
# SPL token dependencies
anchorpy>=0.18.0
# For compatibility with existing code
spl==0.0.1.dev0